Revision: 16582

+++ SequoiaDB/engine/bar/barBkupLogger.cpp
@@ -1360,7 +1360,7 @@ namespace engine
          }
       }
 
-      if ( pLobData->getDataSz() >
+      if ( (UINT64)pLobData->getDataSz() >
            ((UINT64)BAR_THINCOPY_THRESHOLD_SIZE << 20 ) )
       {
          FLOAT64 ratio = (FLOAT64)pLobSU->getSMEMgr()->totalFree() /
@@ -1371,7 +1371,7 @@ namespace engine
          }
       }
 
-      while ( _curOffset < pLobData->getFileSz() )
+      while ( _curOffset < (UINT64)pLobData->getFileSz() )
       {
          if ( cb->isInterrupted() )
          {
